Fix cmd/main_test.go build tag and consolidate test helpers
cmd/main_test.go previously used //go:build client && server, meaning
it was never compiled (and therefore never run) in CI, which passes
-tags=client or -tags=server individually, never both simultaneously.

Fix this by changing the tag to //go:build client || server. The three
test cases that referenced originCmd (a server-only symbol) are replaced
with equivalent cases using the config and config get subcommands, which
are available in both binaries. This preserves the intent of the test —
verifying that --version is intercepted at every level of the command
tree — without requiring any server-only symbols.

cmd/test_helpers_test.go is deleted. It was introduced by the binary
split commit solely to give getPelicanBinary, TestMain, and their
associated vars a home that would compile under client || server while
main_test.go carried the broken client && server tag. With that tag
fixed, there is no reason for a separate file; the functions are moved
back into main_test.go where they originally lived.
